+++ RELEASES.md
@@ -12,6 +12,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-library: added composite template plead_transforming_composite
 * prov-template-compiler: documentation for composite templates
 * prov-template-compiler: forms for composite templates (work in progress)
+* prov-template-compiler: composite templates conversion to csv
 
 
 `2.0.5`